Move `ScrollViewOffsetAnimator` to a new file
